Scapy compatibility with Decimal

+++ TMLib/TMdict.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 
 from . import Definitions as TMdef
 
@@ -101,7 +102,7 @@ class GlobalRWdict(dict):
 
         :param timestamp_shift: Timestamp shift.
         """
-        self[TMdef.ATTACK]['timestamp_shift'] = timestamp_shift
+        self[TMdef.ATTACK]['timestamp_shift'] = Decimal(timestamp_shift)
 
 
     def to_mss_exceptions(self, ip_from):
@@ -189,13 +190,13 @@ class GlobalRWdict(dict):
         if not ip_dict:
             ip_dict = {}
             delay_dict[ip_from] = ip_dict
-        ip_dict[ip_to] = float(avg_delay)
+        ip_dict[ip_to] = Decimal(avg_delay)
 
         ip_dict = delay_dict.get(ip_to)
         if not ip_dict:
             ip_dict = {}
             delay_dict[ip_to] = ip_dict
-        ip_dict[ip_from] = float(avg_delay)
+        ip_dict[ip_from] = Decimal(avg_delay)
 
 
     def to_timestamp_random_delay_threshold_map(self, ip, threshold):
@@ -205,7 +206,7 @@ class GlobalRWdict(dict):
         :param ip: ip address, string
         :param threshold: threshold, float or string
         """
-        self[TMdef.ATTACK]['timestamp_delay_map'][ip] = float(threshold)
+        self[TMdef.ATTACK]['timestamp_delay_map'][ip] = Decimal(threshold)
 
 
     def to_timestamp_random_delay_set(self, ip):
